many changes

+++ TrabajoPractico3/src/ort/edu/tp1/trabajopractico3/ejercicio4/Ticket.java
@@ -14,17 +14,20 @@ public class Ticket implements Imprimible{
 		this.cliente = cliente;
 		this.electrodomesticosComprados = electrodomesticosComprados;
 	}
-
+	
 	@Override
 	public void imprimir() {
 		if(electrodomesticosComprados.size()>0) {
-		int total = 0;
+		double total = 0;
 		System.out.println("Ticket de venta");
 		System.out.println("Articulos:");
+		
 		for(Electrodomestico elec : electrodomesticosComprados) {
 			elec.imprimir();	
 			total += elec.getPrecio();
 		}
+		
+				
 		System.out.println("Total: $" + total);
 		System.out.println("Cliente: ");
 		System.out.println("Cuil: " + cliente.getCuil());
